fix: test cleanup and reorg
